# apps/runner/pkg/api/docs/swagger.yaml
definitions:
...
  CreateBoxDTO:
    properties:
      artifactRef:
        type: string
      authToken:
        type: string
      cpuQuota:
        minimum: 1
        type: integer
      entrypoint:
        items:
          type: string
        type: array
      env:
        additionalProperties:
          type: string
        type: object
      fromVolumeId:
        type: string
      gpuQuota:
        minimum: 0
        type: integer
      id:
        type: string
      memoryQuota:
        minimum: 1
        type: integer
      metadata:
        additionalProperties:
          type: string
        type: object
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      organizationId:
        description: Nullable for backward compatibility
        type: string
      osUser:
        type: string
      otelEndpoint:
        type: string
      regionId:
        type: string
      registry:
        $ref: '#/definitions/RegistryDTO'
      skipStart:
        type: boolean
      storageQuota:
        minimum: 1
        type: integer
      userId:
        type: string
      volumes:
        items:
          $ref: '#/definitions/dto.VolumeDTO'
        type: array
    required:
      - artifactRef
      - id
      - osUser
      - userId
    type: object
  ErrorResponse:
    description: Error response
    properties:
      code:
        example: BAD_REQUEST
        type: string
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: '2023-01-01T12:00:00Z'
        type: string
    required:
      - message
      - path
      - statusCode
      - timestamp
    type: object
...
  IsRecoverableDTO:
    properties:
      errorReason:
        type: string
    required:
      - errorReason
    type: object
  IsRecoverableResponse:
    properties:
      recoverable:
        type: boolean
    type: object
...
  RecoverBoxDTO:
    properties:
      backupErrorReason:
        type: string
      cpuQuota:
        minimum: 1
        type: integer
      env:
        additionalProperties:
          type: string
        type: object
      errorReason:
        type: string
      fromVolumeId:
        type: string
      gpuQuota:
        minimum: 0
        type: integer
      memoryQuota:
        minimum: 1
        type: integer
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      osUser:
        type: string
      snapshot:
        type: string
      storageQuota:
        minimum: 1
        type: integer
      userId:
        type: string
      volumes:
        items:
          $ref: '#/definitions/dto.VolumeDTO'
        type: array
    required:
      - errorReason
      - osUser
      - userId
    type: object
  RegistryDTO:
    properties:
      password:
        type: string
      project:
        type: string
      url:
        type: string
      username:
        type: string
    required:
      - url
    type: object
  ResizeBoxDTO:
    properties:
      cpu:
        minimum: 1
        type: integer
      disk:
        minimum: 1
        type: integer
      gpu:
        minimum: 0
        type: integer
      memory:
        minimum: 1
        type: integer
    type: object
...
  BoxInfoResponse:
    properties:
      backupError:
        type: string
      backupSnapshot:
        type: string
      backupState:
        $ref: '#/definitions/enums.BackupState'
      daemonVersion:
        type: string
      state:
        $ref: '#/definitions/enums.BoxState'
    type: object
  StartBoxResponse:
    properties:
      daemonVersion:
        type: string
    type: object
  StopBoxDTO:
    properties:
      force:
        type: boolean
    type: object
...
  UpdateNetworkSettingsDTO:
    properties:
      networkAllowList:
        type: string
      networkBlockAll:
        type: boolean
      networkLimitEgress:
        type: boolean
    type: object
  dto.VolumeDTO:
    properties:
      mountPath:
        type: string
      subpath:
        type: string
      volumeId:
        type: string
    type: object
  enums.BackupState:
    enum:
      - NONE
      - PENDING
      - IN_PROGRESS
      - COMPLETED
      - FAILED
    type: string
    x-enum-varnames:
      - BackupStateNone
      - BackupStatePending
      - BackupStateInProgress
      - BackupStateCompleted
      - BackupStateFailed
  enums.BoxState:
    enum:
      - creating
      - restoring
      - destroyed
      - destroying
      - started
      - stopped
      - starting
      - stopping
      - resizing
      - error
      - unknown
      - pulling_artifact
    type: string
    x-enum-varnames:
      - BoxStateCreating
      - BoxStateRestoring
      - BoxStateDestroyed
      - BoxStateDestroying
      - BoxStateStarted
      - BoxStateStopped
      - BoxStateStarting
      - BoxStateStopping
      - BoxStateResizing
      - BoxStateError
      - BoxStateUnknown
      - BoxStatePullingArtifact
info:
  contact: {}
  description: BoxLite Runner API
  title: BoxLite Runner API
  version: v0.0.0-dev
paths:
...
  /boxes:
    post:
      description: Create a box
      operationId: Create
      parameters:
        - description: Create box
          in: body
          name: box
          required: true
          schema:
            $ref: '#/definitions/CreateBoxDTO'
      produces:
        - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/StartBoxResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create a box
      tags:
        - box
  /boxes/{boxId}:
    get:
      description: Get box info
      operationId: Info
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Box info
          schema:
            $ref: '#/definitions/BoxInfoResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get box info
      tags:
        - box
...
  /boxes/{boxId}/destroy:
    post:
      description: Destroy box
      operationId: Destroy
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Box destroyed
          schema:
            type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Destroy box
      tags:
        - box
  /boxes/{boxId}/is-recoverable:
    post:
      consumes:
        - application/json
      description: Check if the sandbox's error reason indicates a recoverable error
      operationId: IsRecoverable
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Error reason to check
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/IsRecoverableDTO'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/IsRecoverableResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Check if box error is recoverable
      tags:
        - box
  /boxes/{boxId}/network-settings:
    get:
      description: Get box network settings
      operationId: GetNetworkSettings
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Network settings
          schema:
            $ref: '#/definitions/UpdateNetworkSettingsDTO'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get box network settings
      tags:
        - box
    post:
      description: Update box network settings
      operationId: UpdateNetworkSettings
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Update network settings
          in: body
          name: box
          required: true
          schema:
            $ref: '#/definitions/UpdateNetworkSettingsDTO'
      produces:
        - application/json
      responses:
        '200':
          description: Network settings updated
          schema:
            type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update box network settings
      tags:
        - box
  /boxes/{boxId}/recover:
    post:
      consumes:
        - application/json
      description: Recover box from error state using specified recovery type
      operationId: Recover
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Recovery parameters
          in: body
          name: recovery
          required: true
          schema:
            $ref: '#/definitions/RecoverBoxDTO'
      produces:
        - application/json
      responses:
        '200':
          description: Box recovered
          schema:
            type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Recover box from error state
      tags:
        - box
  /boxes/{boxId}/resize:
    post:
      description: Resize box
      operationId: Resize
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Resize box
          in: body
          name: box
          required: true
          schema:
            $ref: '#/definitions/ResizeBoxDTO'
      produces:
        - application/json
      responses:
        '200':
          description: Box resized
          schema:
            type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Resize box
      tags:
        - box
  /boxes/{boxId}/start:
    post:
      description: Start box
      operationId: Start
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Metadata
          in: body
          name: metadata
          schema:
            type: object
        - description: Auth token
          in: query
          name: token
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: Box started
          schema:
            $ref: '#/definitions/StartBoxResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Start box
      tags:
        - box
  /boxes/{boxId}/stop:
    post:
      description: Stop box
      operationId: Stop
      parameters:
        - description: Box ID
          in: path
          name: boxId
          required: true
          type: string
        - description: Stop box
          in: body
          name: box
          schema:
            $ref: '#/definitions/StopBoxDTO'
      produces:
        - application/json
      responses:
        '200':
          description: Box stopped
          schema:
            type: string
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '401':
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Stop box
      tags:
        - box
...
security:
  - Bearer: []
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and an API token.
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'